% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addCustomAnnotation}
\alias{addCustomAnnotation}
\title{Add a Custom Annotation Column}
\usage{
addCustomAnnotation(spatialExp, columnName, values)
}
\arguments{
\item{spatialExp}{A \code{SpatialExperiment} to annotate.}

\item{columnName}{Name of the new annotation column.}

\item{values}{Vector of annotation values (length must match \code{ncol(spatialExp)}).}
}
\value{
The updated \code{SpatialExperiment} with the new annotation column.
}
\description{
Appends a new column to \code{colData} with user-provided values. This can be
used to add metadata for downstream analyses or visualization.
}
