% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addResultToObject}
\alias{addResultToObject}
\title{Add results to object colData}
\usage{
addResultToObject(spatial_obj, result)
}
\arguments{
\item{spatial_obj}{SpatialExperiment}

\item{result}{deconvolution result, rows = spots, columns = cell types}
}
\description{
Add results to object colData
}
