% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{anndata_to_spatialexperiment}
\alias{anndata_to_spatialexperiment}
\title{Convert SpatialExperiment to AnnData}
\usage{
anndata_to_spatialexperiment()
}
\description{
Convert SpatialExperiment to AnnData
}
