% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{antimode_cutoff}
\alias{antimode_cutoff}
\title{Determine threshold for celltype presence based on antimode of celltype density}
\usage{
antimode_cutoff(m, method)
}
\arguments{
\item{m}{A SpatialExperiment Object}
}
\value{
A vector with celltype specific cutoff values
}
\description{
Determine threshold for celltype presence based on antimode of celltype density
}
