% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{attachToken}
\alias{attachToken}
\title{Attach method token to deconvolution result}
\usage{
attachToken(deconvolution, token = "deconv")
}
\arguments{
\item{deconvolution}{deconvolution result as matrix}

\item{token}{method name or custom token}
}
\value{
deconvolution result with renamed celltypes
}
\description{
Rename Celltypes of deconvolution result and add method token
}
