% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunedeconv.R
\name{build_model_immunedeconv}
\alias{build_model_immunedeconv}
\title{No Signature for Immunedeconv}
\usage{
build_model_immunedeconv()
}
\description{
First-generation immunedeconv methods do not require a custom signature, so
this function returns \code{NULL} and you should call \code{deconvolute()} with the
desired immunedeconv method (see \code{spacedeconv::deconvolution_methods}).
}
