% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omnideconv.R
\name{build_model_omnideconv}
\alias{build_model_omnideconv}
\title{Build Reference}
\usage{
build_model_omnideconv(
  single_cell_obj,
  cell_type_col = "cell_ontology_class",
  method = NULL,
  spatial_obj = NULL,
  batch_id_col = NULL,
  markers = NULL,
  assay_sc = "counts",
  assay_sp = "counts",
  ...
)
}
\arguments{
\item{single_cell_obj}{SingleCellExperiment}

\item{cell_type_col}{column containing cell type annotation}

\item{method}{select deconvolution method}

\item{spatial_obj}{SpatialExperiment}

\item{batch_id_col}{column of single_cell_obj containing batch_ids}

\item{markers}{List of marker genes, only used for bseqsc}

\item{assay_sc}{single cell assay to use}

\item{assay_sp}{spatial assay to use}

\item{...}{additional parameters passed to the selected algorithm}
}
\value{
Gene Expression signature or NULL, if methods don't export signatures
}
\description{
Build Reference
}
