% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization_new.R
\name{cell_pair_presence}
\alias{cell_pair_presence}
\title{Calculate Colocalization (Beta)}
\usage{
cell_pair_presence(
  spe,
  method = NULL,
  distance = 0,
  cell_type_A = NULL,
  cell_type_B = NULL,
  niter = 100,
  presence_matrix = NULL,
  threshold = NULL
)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{method}{presence method}

\item{distance}{distance in spot diameter}

\item{cell_type_A}{celltype 1}

\item{cell_type_B}{celltype 2}

\item{niter}{permutations}

\item{presence_matrix}{presence matrix if available}

\item{threshold}{threshold for custom presence determination}
}
\value{
presence/absence for each cell type
}
\description{
Calculate Colocalization (Beta)
}
