% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_datatype}
\alias{check_datatype}
\title{Convert assays to sparse Matrices}
\usage{
check_datatype(spe, assay = "counts")
}
\arguments{
\item{spe}{SpatialExperiment}

\item{assay}{assay to use}
}
\description{
Convert assays to sparse Matrices
}
