% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkCol}
\alias{checkCol}
\title{Check if Column exists in object}
\usage{
checkCol(object, column)
}
\arguments{
\item{object}{SingleCellExperiment or SpatialExperiment}

\item{column}{column name to check for existence}
}
\value{
if column exists in object
}
\description{
Check if Column exists in object
}
