% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkENSEMBL}
\alias{checkENSEMBL}
\title{Check for ENSEBL IDs}
\usage{
checkENSEMBL(names)
}
\arguments{
\item{names}{vector of rownames}
}
\value{
TRUE if all are ensembl
}
\description{
Check for ENSEBL IDs
}
