% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkRowColumn}
\alias{checkRowColumn}
\title{Check Rowname/Colname Presence}
\usage{
checkRowColumn(object)
}
\arguments{
\item{object}{SingleCellExperiment or SpatialExperiment}
}
\value{
boolean, TRUE if one of rownames/colnames is NULL
}
\description{
Check for Rowname and Column Name existence in expression objects
}
