% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{clean_genes_and_cells}
\alias{clean_genes_and_cells}
\title{Clean Genes and Cells with Metacell2}
\usage{
clean_genes_and_cells(
  anndata,
  properly_sampled_min_cell_total = 800,
  properly_sampled_max_cell_total = 8000,
  properly_sampled_max_excluded_genes_fraction = 0.1,
  exclude_genes = "",
  exclude_gene_patterns = "",
  seed = 123456
)
}
\arguments{
\item{anndata}{anndata object containing Single Cell Expression Data}

\item{properly_sampled_min_cell_total}{testtesttest}

\item{properly_sampled_max_cell_total}{testtesttest}

\item{properly_sampled_max_excluded_genes_fraction}{testtesttest}

\item{exclude_genes}{test}

\item{exclude_gene_patterns}{test}

\item{seed}{seed}
}
\description{
Clean Genes and Cells with Metacell2
}
