% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R, R/colocalization_new.R
\name{coloc_avoid}
\alias{coloc_avoid}
\title{Colocalization of two celltypes}
\usage{
coloc_avoid(A, B)

coloc_avoid(A, B)
}
\arguments{
\item{A}{presence of cell type}

\item{B}{presence of cell type}
}
\description{
Colocalization of two celltypes

Colocalization of two celltypes
}
