% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{compare_signatures}
\alias{compare_signatures}
\title{Compare Signatures}
\usage{
compare_signatures(signature1, signature2)
}
\arguments{
\item{signature1}{Signature matrix.}

\item{signature2}{Signature matrix.}
}
\description{
Creates a scatterplot comparing two signature matrices across shared genes.
}
