% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_conversion.R
\name{convert_human_to_mouse}
\alias{convert_human_to_mouse}
\title{Map Human to Mouse Gene Symbols}
\usage{
convert_human_to_mouse(humangenes)
}
\arguments{
\item{humangenes}{Vector of human gene symbols to convert.}
}
\description{
Converts human gene symbols to mouse gene symbols using
\code{Orthology.eg.db} mappings.
}
