% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_conversion.R
\name{convert_mouse_to_human}
\alias{convert_mouse_to_human}
\title{Map Mouse to Human Gene Symbols}
\usage{
convert_mouse_to_human(mousegenes)
}
\arguments{
\item{mousegenes}{Vector of mouse gene symbols to convert.}
}
\description{
Converts mouse gene symbols to human gene symbols using
\code{Orthology.eg.db} mappings.
}
