% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convert_to_sce}
\alias{convert_to_sce}
\title{Convert Input Files to SingleCellExperiment}
\usage{
convert_to_sce(obj)
}
\arguments{
\item{obj}{object provided by the user, will be converted to sce}
}
\description{
Convert Input Files to SingleCellExperiment
}
