% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunedeconv.R
\name{convertImmunedeconvMatrix}
\alias{convertImmunedeconvMatrix}
\title{Convert Immunedeconv Matrix to match SpaceDeconv Format}
\usage{
convertImmunedeconvMatrix(deconvResult)
}
\arguments{
\item{deconvResult}{immunedeconv result matrix}
}
\value{
transformed matrix where cell types are columns
}
\description{
Convert Immunedeconv Matrix to match SpaceDeconv Format
}
