% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialDWLS.R
\name{doGiottoWorkflow}
\alias{doGiottoWorkflow}
\title{Perform Giotto Workflow of Normalization, HVG, Dimension Reduction and Clustering}
\usage{
doGiottoWorkflow(
  obj,
  calculateHVG = TRUE,
  dim_method = "pca",
  cluster_method = "leiden",
  show_plot = FALSE,
  return_plot = FALSE,
  save_plot = FALSE,
  ...
)
}
\arguments{
\item{obj}{Giotto Object}

\item{calculateHVG}{Wheter to calculate HVG}

\item{dim_method}{dimension reduction method to use: pca, tsne, umap}

\item{cluster_method}{cluster method to use c(leiden, kmeans, hclust, louvain)}

\item{show_plot}{show plot}

\item{return_plot}{return plot}

\item{save_plot}{save plot}

\item{...}{additional parameters}
}
\description{
Perform Giotto Workflow of Normalization, HVG, Dimension Reduction and Clustering
}
