% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{extract_forbidden_from_modules}
\alias{extract_forbidden_from_modules}
\title{extract forbidden genes from gene modules}
\usage{
extract_forbidden_from_modules(clean, forbidden_modules)
}
\arguments{
\item{clean}{anndata object}

\item{forbidden_modules}{test}
}
\description{
extract forbidden genes from gene modules
}
