% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{filter_sample_id}
\alias{filter_sample_id}
\title{Filter by Sample ID}
\usage{
filter_sample_id(spe, sample_id)
}
\arguments{
\item{spe}{\code{SpatialExperiment}.}

\item{sample_id}{Sample ID to keep.}
}
\description{
Subsets a \code{SpatialExperiment} to a single \code{sample_id}.
}
