% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_hex_polygon}
\alias{get_hex_polygon}
\title{Build Hex Polygon Geometry}
\usage{
get_hex_polygon(x, y, dist, offset_rotation = FALSE)
}
\arguments{
\item{x}{coordinate}

\item{y}{coordinate}

\item{dist}{distance of hexagons}

\item{offset_rotation}{correct hex orientation for rotated visium image}
}
\description{
Build Hex Polygon Geometry
}
