% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R, R/colocalization_new.R
\name{get_iniche}
\alias{get_iniche}
\title{Calculate Iniche recursivley}
\usage{
get_iniche(df, coordinates, distance)

get_iniche(df, coordinates, distance)
}
\arguments{
\item{df}{dataframe with array_row and array_col}

\item{coordinates}{vector of (row, columns)}

\item{distance}{spot distance}
}
\value{
list of spot ids of iniche

list of spot ids of iniche
}
\description{
Calculate Iniche recursivley

Calculate Iniche recursivley
}
