% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripleys.R
\name{get_largest_r_and_iso}
\alias{get_largest_r_and_iso}
\title{Largest radius and isotropic value}
\usage{
get_largest_r_and_iso(k_functions)
}
\arguments{
\item{k_functions}{a list with named ripley´s k statistics}
}
\value{
largest radius and isotropic value in a list of k statistics
}
\description{
Largest radius and isotropic value
}
