% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lr.R
\name{get_ligand_expression}
\alias{get_ligand_expression}
\title{Get Ligand Expression}
\usage{
get_ligand_expression(gene_pair, cpm_df)
}
\arguments{
\item{gene_pair}{A string representing a gene or a gene complex. Gene complexes are indicated by genes separated by underscores.}

\item{cpm_df}{A dataframe containing gene expression data, where row names are gene names and columns represent different samples or spots.}
}
\value{
A numeric vector containing the expression values of the specified ligand across the samples or spots in \code{cpm_df}. If the ligand is a gene complex, it returns the mean expression value of the component genes. If any component gene is missing or has zero expression, it returns a vector of zeros.
}
\description{
This function computes the expression values of a specified ligand (gene or gene complex) from a given expression dataset.
}
