% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{get_mostAbundantInCluster}
\alias{get_mostAbundantInCluster}
\title{Get Most Abundant cell types per cluster}
\usage{
get_mostAbundantInCluster(spe, method = "mean", k = 3)
}
\arguments{
\item{spe}{SpatialExperiment with deconvolution results}

\item{method}{how to calculate (sum, mean)}

\item{k}{return top k most abundant cell types}
}
\description{
Get Most Abundant cell types per cluster
}
