% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_polygon_geometry}
\alias{get_polygon_geometry}
\title{Build Full Hexagon Set for provided spots}
\usage{
get_polygon_geometry(grid, dist)
}
\arguments{
\item{grid}{Spot coordinates in df format, extracted from SpatialExperiment}

\item{dist}{distance between spots}
}
\description{
Build Full Hexagon Set for provided spots
}
