% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lr.R
\name{get_receptor_expression}
\alias{get_receptor_expression}
\title{Get Receptor Expression}
\usage{
get_receptor_expression(gene_pair, cpm_df)
}
\arguments{
\item{gene_pair}{A string representing a gene or a gene complex, where gene complexes are indicated by genes separated by underscores, and the receptor gene is specified after a "~".}

\item{cpm_df}{A dataframe containing gene expression data, with row names as gene names and columns as different samples or spots.}
}
\value{
A numeric vector of the expression values of the specified receptor across the samples or spots in \code{cpm_df}. If the receptor is a gene complex, the function returns the mean expression value of the component genes. Missing component genes give zero for all spots; zero expression in a component gives zero only for the affected spots.
}
\description{
This function calculates the expression values of a specified receptor (gene or gene complex) from a given expression dataset.
}
