% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_results_from_object}
\alias{get_results_from_object}
\title{get deconvolution results from object}
\usage{
get_results_from_object(spatial_obj)
}
\arguments{
\item{spatial_obj}{SpatialExperiment}
}
\description{
get deconvolution results from object
}
