% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{get_spot_coordinates}
\alias{get_spot_coordinates}
\title{Get coordinates of spot id}
\usage{
get_spot_coordinates(df, spotid)
}
\arguments{
\item{df}{colData dataframe}

\item{spotid}{spotid}
}
\description{
Get coordinates of spot id
}
