% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMatricesFromSCE}
\alias{getMatricesFromSCE}
\title{get matrices from SingleCellExperiment}
\usage{
getMatricesFromSCE(single_cell_object, cell_type_col = "cell_ontology_class")
}
\arguments{
\item{single_cell_object}{SingleCellExperiment}

\item{cell_type_col}{column containing the cell type}
}
\value{
list containing the expression matrix and cell type annotation vector
}
\description{
get matrices from SingleCellExperiment
}
