% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{make_baseplot}
\alias{make_baseplot}
\title{Render spatial hex plot}
\usage{
make_baseplot(
  spe,
  df,
  to_plot,
  palette = "Rocket",
  transform_scale = NULL,
  sample_id = "sample01",
  reverse_palette = FALSE,
  image_id = "lowres",
  show_image = TRUE,
  discrete = FALSE,
  offset_rotation = FALSE,
  spot_size = 1,
  limits = NULL
)
}
\arguments{
\item{spe}{SpatialExperiment with deconvolution results}

\item{df}{containing the annotation to be plotted}

\item{to_plot}{column of df to plot}

\item{palette}{colorspace palette (sequential)}

\item{transform_scale}{data transform_scaleation to use, "log"}

\item{sample_id}{sample of the SpatialExperiment to be plotted}

\item{reverse_palette}{reverse color palette}

\item{image_id}{image id for background image}

\item{show_image}{whether to show the spatial image}

\item{discrete}{should the color scale be discrete? Defaut = FALSE}

\item{offset_rotation}{correct hex orientation for rotated visium image}

\item{spot_size}{increase (>1) or decrease (<1) the hex size}

\item{limits}{vector of color scale limits}
}
\description{
Render spatial hex plot
}
