% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_cells_per_spot}
\alias{plot_cells_per_spot}
\title{Plot Cells per Spot}
\usage{
plot_cells_per_spot(spatial_obj, treshold = 0)
}
\arguments{
\item{spatial_obj}{SpatialExperiment containing deconvolution results}

\item{treshold}{treshold for presence/absence, single value or vector of length nrow(spatial_obj)}
}
\description{
Plot Cells per Spot
}
