% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_celltype}
\alias{plot_celltype}
\title{Plot Spatial Object}
\usage{
plot_celltype(spatial_obj, sample = "sample01", cell_type = NULL)
}
\arguments{
\item{spatial_obj}{SpatialExperiment containing deconvolution results}

\item{sample}{sample to plot}

\item{cell_type}{Cell Type to plot}
}
\description{
Plot Spatial Object
}
