% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_fade_map}
\alias{plot_fade_map}
\title{Function to plot selected cell types with transparency scaling + color legend}
\usage{
plot_fade_map(
  spe,
  cell_types,
  colors = NULL,
  transform = "none",
  sample_id = "sample01",
  image_id = "lowres",
  show_image = TRUE,
  background = NULL,
  spot_size = 1,
  limits = NULL,
  save = FALSE,
  path = NULL,
  png_width = 1500,
  png_height = 750
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing deconvolution results}

\item{cell_types}{A vector of manually selected cell types to plot (max 6)}

\item{colors}{A vector of colors for the selected cell types (optional; default colors used if NULL)}

\item{transform}{Data transformation to apply ("none", "log", "log10", "log2", "sqrt")}

\item{sample_id}{The sample ID to plot, default: "sample01"}

\item{image_id}{Which image to plot, default: "lowres"}

\item{show_image}{Logical, whether to display the image, default = TRUE}

\item{background}{Background color}

\item{spot_size}{Increase (>1) or decrease (<1) the hex size}

\item{limits}{Vector of color scale limits}

\item{save}{Set TRUE to save plot}

\item{path}{Directory to save plot, default: ~/spacedeconv}

\item{png_width}{PNG width in pixels when saving}

\item{png_height}{PNG height in pixels when saving}
}
\value{
A ggplot2 object containing the overlayed transparency-based hex plot
}
\description{
Generates a hex plot where each selected cell type is visualized independently,
with color intensity proportional to its abundance.
}
