% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell_plots.R
\name{plot_metacell_abundance}
\alias{plot_metacell_abundance}
\title{Plot metacell abundance score}
\usage{
plot_metacell_abundance(metacell)
}
\arguments{
\item{metacell}{SingleCellExperiment containing metacells}
}
\description{
Plot metacell abundance score
}
