% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell_plots.R
\name{plot_metacell_nCells}
\alias{plot_metacell_nCells}
\title{Plot the number of genes per metacells}
\usage{
plot_metacell_nCells(metacell)
}
\arguments{
\item{metacell}{SingleCellExperiment containing metacells}
}
\description{
Plot the number of genes per metacells
}
