% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_most_abundant}
\alias{plot_most_abundant}
\title{Function to plot deconvolution results}
\usage{
plot_most_abundant(
  spe,
  method = NULL,
  cell_type = NULL,
  remove = NULL,
  palette = "Mako",
  sample_id = "sample01",
  image_id = "lowres",
  reverse_palette = FALSE,
  show_image = TRUE,
  offset_rotation = FALSE,
  spot_size = 1.17,
  font_size = 15
)
}
\arguments{
\item{spe}{deconvolution result in Form of a SpatialExperiment}

\item{cell_type}{one or more celltype to plot, NULL for all}

\item{palette}{colorspace palette (sequential)}

\item{sample_id}{sample id to plot, default: "sample01"}

\item{image_id}{which image to plot, default: "lowres"}

\item{reverse_palette}{reverse color palette}

\item{show_image}{logical, whether to display the image, default = TRUE}

\item{offset_rotation}{correct hex orientation for rotated visium image}

\item{spot_size}{increase (>1) or decrease (<1) the hex size}

\item{font_size}{font size}
}
\value{
plot of cell type fractions
}
\description{
Generate Hex Plot of a SpatialExperiment containing the most abundant cell types
}
