% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripleys.R
\name{plot_ripleys_k}
\alias{plot_ripleys_k}
\title{Plot multiple Ripley´s K statistics}
\usage{
plot_ripleys_k(k_functions)
}
\arguments{
\item{k_functions}{a list with named ripley´s k statistics}
}
\value{
combined ripley´s k plot
}
\description{
Plot multiple Ripley´s K statistics
}
