% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_umi_count}
\alias{plot_umi_count}
\title{Function to plot deconvolution results}
\usage{
plot_umi_count(
  spe,
  palette = "Mako",
  transform_scale = NULL,
  sample_id = "sample01",
  image_id = "lowres",
  reverse_palette = FALSE,
  zoom = TRUE,
  show_image = FALSE,
  background = NULL,
  offset_rotation = FALSE,
  spot_size = 1,
  limits = NULL,
  smooth = FALSE,
  smoothing_factor = 1.5,
  title_size = 30,
  title = NULL,
  font_size = 15,
  legend_size = 20,
  density = TRUE,
  save = FALSE,
  path = NULL,
  png_width = 1500,
  png_height = 750,
  show_legend = TRUE,
  ...
)
}
\arguments{
\item{spe}{deconvolution result in Form of a SpatialExperiment}

\item{palette}{colorspace palette (sequential)}

\item{transform_scale}{data transform_scaleation to use, "log"}

\item{sample_id}{sample id to plot, default: "sample01"}

\item{image_id}{which image to plot, default: "lowres"}

\item{reverse_palette}{reverse color palette}

\item{zoom}{zoom to the available spots}

\item{show_image}{logical, wether to display the image, default = TRUE}

\item{background}{custom background color}

\item{offset_rotation}{correct hex orientation for rotated visium image}

\item{spot_size}{increase (>1) or decrease (<1) the hex size}

\item{limits}{vector of color scale limits}

\item{smooth}{whether to smooth the plot}

\item{smoothing_factor}{kernel size factor (multiples of spot distance)}

\item{title_size}{font size of title}

\item{title}{set a custom title}

\item{font_size}{font size of legend}

\item{legend_size}{legend size in points}

\item{density}{whether to display a density distribution next to the spatial plot}

\item{save}{set TRUE to save plot}

\item{path}{specify directory to save plot, if NULL: saving at ~/spacedeconv}

\item{png_width}{when saving, png width in px}

\item{png_height}{when saving, png height in px}

\item{show_legend}{whether to show the legend}

\item{...}{additional paramters passed to internal functions}
}
\value{
plot of cell type fractions
}
\description{
Generate Hex Plot of a SpatialExperiment containing UMI counts
}
\examples{
data("spatial_data_3")
deconv <- spacedeconv::deconvolute(spatial_data_3, method = "estimate")
plot_umi_count(deconv)
}
