% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{presence}
\alias{presence}
\title{Threshold a matrix}
\usage{
presence(m, trheshold)
}
\arguments{
\item{m}{matrix}

\item{threshold}{if single value is provided the same threshold is use for all celltypes, it is also possible to provide a threshold vector}
}
\description{
Threshold a matrix
}
