% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_info.R
\name{print_info}
\alias{print_info}
\title{Print Dataset Summary}
\usage{
print_info(sce = NULL, spe = NULL, signature = NULL, assay = "counts")
}
\arguments{
\item{sce}{\code{SingleCellExperiment} (optional).}

\item{spe}{\code{SpatialExperiment} (optional).}

\item{signature}{Signature matrix (optional).}

\item{assay}{Assay to use for computations (default: "counts").}
}
\description{
Prints basic QC summaries for \code{SingleCellExperiment}, \code{SpatialExperiment},
and optional signature matrices (counts, zeros, UMI range, overlap).
}
