% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeZeroExpression}
\alias{removeZeroExpression}
\title{Remove Spots with zero expression}
\usage{
removeZeroExpression(object)
}
\arguments{
\item{object}{SummarizedExperiment or any related datatypes}
}
\value{
Expression Object without all zero columns
}
\description{
This function removes spots/columns with zero expression detected. These spots might result in errors during computation
}
