% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{ripleys_key}
\alias{ripleys_key}
\title{Ripley´s K celltype distribution}
\usage{
ripleys_key(spe, cell_type, method)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{cell_type}{celltype of interest}

\item{method}{deconvolution method}
}
\description{
Ripley´s K celltype distribution
}
