% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{save_plot}
\alias{save_plot}
\title{Save Plot to path}
\usage{
save_plot(plot, to_plot, path, png_width, png_height)
}
\arguments{
\item{plot}{ggplot}

\item{to_plot}{celltype name}

\item{path}{where the plot should be stored}

\item{png_width}{when saving, png width in px}

\item{png_height}{when saving, png height in px}
}
\description{
Save Plot to path
}
