% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_cibersort_mat}
\alias{set_cibersort_mat}
\title{Register the Path to the CIBERSORT Signature Matrix}
\usage{
set_cibersort_mat(path)
}
\arguments{
\item{path}{Path to the \code{LM22.txt} signature matrix file.}
}
\description{
Stores the path to the \code{LM22.txt} signature matrix in the spacedeconv configuration
and forwards it to \code{immunedeconv::set_cibersort_mat()} so CIBERSORT-based methods
can access the signature from this R session.
}
\details{
CIBERSORT is only freely available to academic users. A license and matrix can
be obtained from https://cibersort.stanford.edu.
}
