% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{seurat_to_spatialexperiment}
\alias{seurat_to_spatialexperiment}
\title{Convert Seurat to SpatialExperiment}
\usage{
seurat_to_spatialexperiment(seurat)
}
\arguments{
\item{seurat}{Seurat Object}
}
\value{
SpatialExperiment
}
\description{
Convert Seurat to SpatialExperiment
}
