% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_cell_data_1}
\alias{single_cell_data_1}
\title{Example single cell dataset 1 from Wu et al.}
\format{
A SingleCellExperiment with 29733 rows and 5789 columns
}
\source{
Wu SZ, Al-Eryani G, Roden DL, Junankar S et al. A single-cell and
spatially resolved atlas of human breast cancers. Nat Genet 2021 Sep;53(9):1334-1347.
PMID: 34493872
}
\usage{
data("single_cell_data_1")
}
\description{
A single cell expression subset of 5789 cells from sample CID4290. Cell type annotation is
available in three different resolutions.
}
\keyword{datasets}
