% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{smooth_celltype}
\alias{smooth_celltype}
\title{Smooth spot annotation}
\usage{
smooth_celltype(df, spot_distance, smoothing_factor = 1.5, cell_type = NULL)
}
\arguments{
\item{df}{DataFrame containing spot coordinates and the value to be smoothed}

\item{spot_distance}{distance between two spots}

\item{smoothing_factor}{multiplied with spot distance to find close spots}

\item{cell_type}{value to be smoothed}
}
\description{
Smooth spot annotation
}
