% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacedeconv_pkg.R
\docType{package}
\name{spacedeconv_pkg}
\alias{spacedeconv}
\alias{spacedeconv-package}
\alias{spacedeconv_pkg}
\title{spacedeconv}
\description{
spacedeconv Unified R Interface to spatial transcriptomics deconvolution methods
}
\author{
\strong{Maintainer}: Constantin Zackl \email{constantin.zackl@student.uibk.ac.at} (\href{https://orcid.org/0000-0003-1991-6943}{ORCID})

}
