% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_data_3}
\alias{spatial_data_3}
\title{Example spatial dataset 3 from Wu et al.}
\format{
A SpatialExperiment with 36601 genes and 1322 spots
}
\source{
Wu SZ, Al-Eryani G, Roden DL, Junankar S et al. A single-cell and
spatially resolved atlas of human breast cancers. Nat Genet 2021 Sep;53(9):1334-1347.
PMID: 34493872
}
\usage{
spatial_data_3
}
\description{
A SpatialExperiment containing a total of 1322 spots from sample CID44971
}
\keyword{datasets}
