% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{spe_to_ad}
\alias{spe_to_ad}
\title{Convert Spatial Experiment to AnnData}
\usage{
spe_to_ad(spe, assay = "counts")
}
\arguments{
\item{spe}{Spatial experiment}

\item{assay}{assay to use}
}
\description{
Convert Spatial Experiment to AnnData
}
