% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{topfeat}
\alias{topfeat}
\title{Get Top Features}
\usage{
topfeat(idx, scores, topn)
}
\arguments{
\item{idx}{ids}

\item{scores}{scores}

\item{topn}{number of features}
}
\description{
Get Top Features
}
