% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addCustomAnnotation}
\alias{addCustomAnnotation}
\title{Extend a SpatialExperiment with a New Annotation Column}
\usage{
addCustomAnnotation(spatialExp, columnName, values)
}
\arguments{
\item{spatialExp}{A \code{SpatialExperiment} object that will be extended with the new annotation.}

\item{columnName}{A \code{character} string that names the new annotation column to be added.}

\item{values}{A numeric or character vector containing the annotation values to be assigned to each spot}
}
\value{
The function returns an updated \code{SpatialExperiment} object that includes the newly added annotation
column within its \code{colData}.
}
\description{
This function is designed to annotate a \code{SpatialExperiment} object by appending a custom annotation
column to its \code{colData}. This feature allows users to integrate additional metadata or experimental
results that can be useful for subsequent analyses or visualizations.
}
