% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addCustomAnnotation}
\alias{addCustomAnnotation}
\title{Add Custom Annotation to a SpatialExperiment}
\usage{
addCustomAnnotation(spatialExp, columnName, values)
}
\arguments{
\item{spatialExp}{A \code{SpatialExperiment} object to which the annotation will be added.}

\item{columnName}{A \code{character} string representing the name of the new annotation column.}

\item{values}{A vector of values for the annotation. The length of this vector must match
the number of columns in the \code{SpatialExperiment} object.}
}
\value{
A \code{SpatialExperiment} object with the new annotation added to its \code{colData}.
}
\description{
This function adds a new custom annotation column to the \code{colData} of a
\code{SpatialExperiment} object.
}
