% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_results}
\alias{aggregate_results}
\title{Aggregate Deconvolution results}
\usage{
aggregate_results(
  spatial_obj = NULL,
  cell_type_1 = NULL,
  cell_type_2 = NULL,
  name = NULL,
  remove = FALSE
)
}
\arguments{
\item{spatial_obj}{SpatialExperiment containing deconvolution results}

\item{cell_type_1}{cell type to aggregate, including the method "method_celltype"}

\item{cell_type_2}{cell type to aggregate, including the method "method_celltype"}

\item{name}{new name for aggregation}

\item{remove}{logical, remove provided cell types and just keep the aggregation}
}
\value{
SpatialObject containing aggregation of provided cell types
}
\description{
Aggregate deconvolution results by merging result columns in the spatial object
}
