% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{anndata_to_spatialexperiment}
\alias{anndata_to_spatialexperiment}
\title{Convert AnnData to SpatialExperiment}
\usage{
anndata_to_spatialexperiment(ad)
}
\arguments{
\item{ad}{anndata object}
}
\description{
Convert AnnData to SpatialExperiment
}
