% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{annotate_spots}
\alias{annotate_spots}
\title{Annotate spots}
\usage{
annotate_spots(
  spe,
  spots,
  value_pos = TRUE,
  value_neg = FALSE,
  name = "annotation"
)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{spots}{list of spots to annotate}

\item{value_pos}{positive value}

\item{value_neg}{negative value}

\item{name}{name of the annotation}
}
\value{
SpatialExperiment containing annotation
}
\description{
Annotate spots
}
