% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization_new.R
\name{antimode_cutoff}
\alias{antimode_cutoff}
\title{Determine threshold for celltype presence based on antimode of celltype density}
\usage{
antimode_cutoff(spe, method)
}
\arguments{
\item{spe}{A SpatialExperiment Object}

\item{method}{deconvolution method}
}
\value{
A vector with celltype specific cutoff values
}
\description{
Determine threshold for celltype presence based on antimode of celltype density
}
