% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_results}
\alias{available_results}
\title{Check wich deconvolutionr results are available in a SpatialExperiment object}
\usage{
available_results(deconv)
}
\arguments{
\item{deconv}{SpatialExperiment}
}
\description{
Check wich deconvolutionr results are available in a SpatialExperiment object
}
