% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_algorithms.R
\name{build_and_deconvolute}
\alias{build_and_deconvolute}
\title{Build Model and Deconvolute in one step}
\usage{
build_and_deconvolute(
  single_cell_obj,
  spatial_obj,
  method = NULL,
  cell_type_col = "cell_ontology_class",
  batch_id_col = NULL,
  assay_sc = "counts",
  assay_sp = "counts",
  return_object = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{single_cell_obj}{Single Cell Object containing reference data to build the model}

\item{spatial_obj}{SpatialExperiment to be deconvoluted}

\item{method}{deconvolution method}

\item{cell_type_col}{column of single_cell_obj containing cell type information}

\item{batch_id_col}{column of SpatialObject containing batch_id information}

\item{assay_sc}{the assay of the single cell object to use, default = "counts"}

\item{assay_sp}{the assay of the spatialExperiment to use, default = "counts"}

\item{return_object}{if true return anotation spatialExperiment, if false return table}

\item{verbose}{output more information}

\item{...}{further parameters passed to the selected function}
}
\description{
Build Model and Deconvolute in one step
}
