% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_algorithms.R
\name{build_model}
\alias{build_model}
\title{Build Reference}
\usage{
build_model(
  single_cell_object,
  cell_type_col = "cell_ontology_class",
  method = NULL,
  verbose = FALSE,
  spatial_object = NULL,
  batch_id_col = NULL,
  ...
)
}
\arguments{
\item{single_cell_object}{Single Cell Expression Object}

\item{cell_type_col}{Name of the anntotation column containing cell type information}

\item{method}{signature calculation method}

\item{verbose}{Display more information on console}

\item{spatial_object}{SpatialExperiment, required for SPOTlight}

\item{batch_id_col}{column of singleCellExperiment containing batch ids}

\item{...}{additional parameters passed to the functions}
}
\description{
Build Reference
}
