% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD.R
\name{build_model_card}
\alias{build_model_card}
\title{No signature calculated, just call the deconvolute method}
\usage{
build_model_card()
}
\description{
No signature calculated, just call the deconvolute method
}
