% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell2location.R
\name{build_model_cell2location}
\alias{build_model_cell2location}
\title{Build Model Cell2location}
\usage{
build_model_cell2location(
  single_cell_obj,
  epochs = 250,
  assay_sc = "counts",
  batch_id_col = "sample_id",
  cell_type_col = "celltype_major",
  cell_count_cutoff = 5,
  cell_percentage_cutoff = 0.03,
  nonz_mean_cutoff = 1.12,
  gpu = TRUE
)
}
\arguments{
\item{single_cell_obj}{SingleCellExperiment containing reference expression}

\item{epochs}{number of epochs to train the network}

\item{assay_sc}{assay to extract}

\item{batch_id_col}{batch_id_col column}

\item{cell_type_col}{column of single_cell_obj containing cell type labels}

\item{cell_count_cutoff}{cell2location parameter}

\item{cell_percentage_cutoff}{cell2location parameter}

\item{nonz_mean_cutoff}{cell2location parameter}

\item{gpu}{whether to train on GPU}
}
\description{
Build Model Cell2location
}
