% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunedeconv.R
\name{build_model_immunedeconv}
\alias{build_model_immunedeconv}
\title{No signature calculated, just call the deconvolute method}
\usage{
build_model_immunedeconv()
}
\description{
No signature calculated, just call the deconvolute method
}
