% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTD.R
\name{build_model_rctd}
\alias{build_model_rctd}
\title{RCTD builds the model internally, please just use the deconvolute method}
\usage{
build_model_rctd()
}
\description{
RCTD builds the model internally, please just use the deconvolute method
}
