% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialDWLS.R
\name{build_model_spatial_dwls}
\alias{build_model_spatial_dwls}
\title{Build Model Spatial DWLS}
\usage{
build_model_spatial_dwls(
  sce,
  assay_sc = "counts",
  marker_method = "scran",
  cell_type_col = "cell_ontology_class",
  ...
)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{assay_sc}{Single Cell Object assay to use}

\item{marker_method}{provide list of marker genes or method to calculate markers (scran, gini, mast)}

\item{cell_type_col}{column of sce containing cell type information}

\item{...}{additional paramters}
}
\description{
Build Model Spatial DWLS
}
