% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPOTlight.R
\name{build_model_spotlight}
\alias{build_model_spotlight}
\title{Build Model SPOTlight}
\usage{
build_model_spotlight(
  sce,
  cell_type_col = "cell_ontology_class",
  spe,
  markers = NULL
)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{cell_type_col}{Column where cell type information can be found}

\item{spe}{SpatialExperiment}

\item{markers}{(Optional) Marker Gene DataFrame, if NULL markers will be calculated from 'sce' based on the authors suggestion}
}
\description{
Build Model SPOTlight
}
