% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPOTlight.R
\name{build_model_spotlight}
\alias{build_model_spotlight}
\title{Build Model SPOTlight}
\usage{
build_model_spotlight(
  single_cell_obj,
  cell_type_col = "cell_ontology_class",
  spatial_obj,
  assay_sc = "counts",
  assay_sp = "counts",
  markers = NULL
)
}
\arguments{
\item{single_cell_obj}{SingleCellExperiment}

\item{cell_type_col}{Column where cell type information can be found}

\item{spatial_obj}{SpatialExperiment}

\item{assay_sc}{single cell assay to use}

\item{assay_sp}{spatial assay to use}

\item{markers}{(Optional) Marker Gene DataFrame, if NULL markers will be calculated from 'single_cell_obj' based on the authors suggestion}
}
\description{
Build Model SPOTlight
}
