% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R, R/colocalization_new.R
\name{cell_pair_localization}
\alias{cell_pair_localization}
\title{Calculate Colocalization}
\usage{
cell_pair_localization(
  spe,
  method = NULL,
  distance = 0,
  cell_type_A = NULL,
  cell_type_B = NULL,
  density = FALSE,
  niter = 100,
  presence_matrix = NULL,
  threshold = NULL
)

cell_pair_localization(
  spe,
  method = NULL,
  distance = 0,
  cell_type_A = NULL,
  cell_type_B = NULL,
  density = FALSE,
  niter = 100,
  presence_matrix = NULL,
  threshold = NULL
)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{method}{presence method}

\item{distance}{distance in spot diameter}

\item{cell_type_A}{celltype 1}

\item{cell_type_B}{celltype 2}

\item{density}{logical}

\item{niter}{permutations}

\item{presence_matrix}{presence matrix if available}

\item{threshold}{threshold for custom presence determination}

\item{cell_type_1}{celltype 1}

\item{cell_type_2}{celltype 2}

\item{presence}{presence matrix if available}
}
\value{
statistics and graph

statistics and graph
}
\description{
Calculate Colocalization

Calculate Colocalization
}
