% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cell_pair_localization}
\alias{cell_pair_localization}
\title{Calculate Colocalization}
\usage{
cell_pair_localization(
  spe,
  method = NULL,
  distance = 0,
  cell_type_1 = NULL,
  cell_type_2 = NULL,
  density = FALSE,
  niter = 100
)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{distance}{distance in spot diameter}

\item{density}{logical}
}
\description{
Calculate Colocalization
}
