% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{clean_genes_and_cells}
\alias{clean_genes_and_cells}
\title{Clean Genes and Cells with Metacell2}
\usage{
clean_genes_and_cells(
  anndata,
  properly_sampled_min_cell_total = 800,
  properly_sampled_max_cell_total = 8000,
  properly_sampled_max_excluded_genes_fraction = 0.1,
  exclude_genes = "",
  exclude_gene_patterns = "",
  seed = 123456
)
}
\arguments{
\item{anndata}{anndata object containing Single Cell Expression Data}

\item{properly_sampled_min_cell_total}{exclude all cells with UMI count below this threshold}

\item{properly_sampled_max_cell_total}{exclude all cells with UMI count above this threshold}

\item{properly_sampled_max_excluded_genes_fraction}{exclude all cells whose sum of excluded data divided by the toal is more then the threshold}

\item{exclude_genes}{gene names to exlude}

\item{exclude_gene_patterns}{gene patterns to exclude}

\item{seed}{seed for reproducibilty}
}
\description{
Clean Genes and Cells with Metacell2
}
