% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Cluster Spots}
\usage{
cluster(
  spe,
  method = "kmeans",
  cluster = "expression",
  assay = "counts",
  nclusters = 3,
  ...
)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{method}{clustering method, (kmeans)}

\item{cluster}{what to cluster, c("expression", "deconvolution")}

\item{assay}{assay to use for clustering}

\item{nclusters}{number of clusters, for kmeans}

\item{...}{further arguments passed to the clustering functions}
}
\value{
SpatialExperiment containing cluster information
}
\description{
Cluster spots and add cluster annotation to the spatial object
}
