% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Cluster spacedeconv results}
\usage{
cluster(
  spe,
  method = c("kmeans", "hclust"),
  data = c("deconvolution", "expression", "pathway", "tf"),
  dist_method = c("correlation", "euclidean", "maximum", "manhattan", "canberra",
    "binary", "minkowski"),
  hclust_method = c("complete", "ward.D", "ward.D2", "single", "average", "mcquitty",
    "median", "centroid"),
  nclusters = 3,
  spmethod = NULL,
  pca_dim = seq(1, 30),
  clusres = 0.5,
  ...
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing the data to be clustered.}

\item{method}{A character vector specifying the clustering method to use.
Options are "kmeans" and "hclust". Default is c("kmeans", "hclust").}

\item{data}{A character vector indicating the type of data to cluster.
Options are "deconvolution", "expression", "pathway", or "tf".
Default is c("deconvolution", "expression", "pathway", "tf").}

\item{dist_method}{A character vector specifying the distance measure to be
used if method = "hclust". Default options are "correlation",
"euclidean", "maximum", "manhattan", "canberra", "binary",
"minkowski".}

\item{hclust_method}{A character vector indicating the agglomeration method
to be used if method = "hclust". Default options are "complete",
"ward.D", "ward.D2", "single", "average", "mcquitty", "median",
"centroid".}

\item{nclusters}{An integer specifying the number of clusters to create.
Default is 3.}

\item{spmethod}{A character string indicating the spatial method used for
the clustering. Must be one of "dorothea", "progeny", "expression",
or the name of the deconvolution method used. Default is NULL.}

\item{pca_dim}{An integer vector specifying PCA dimensions to be used for
clustering of expression data. This is used with Seurat::FindNeighbors.
Default is seq(1, 30).}

\item{clusres}{A numeric value for the clustering resolution to be used
with Seurat::FindClusters. Default is 0.5.}

\item{...}{Additional parameters.}
}
\value{
The modified SpatialExperiment object with clustering results.
}
\description{
Performs clustering on data obtained from a SpatialExperiment object using
specified methods. This function allows clustering based on deconvolution
results, expression data, pathway, or transcription factors (TF) analyses.
}
\examples{
# Example usage
data("spatial_data_3")
spe_clustered <- spacedeconv::cluster(spatial_data_3, method = "kmeans", data = "expression")

}
