% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR.R
\name{compute_decoupleR_activites}
\alias{compute_decoupleR_activites}
\title{Compute Pathway activities with decoupleR}
\usage{
compute_decoupleR_activites(
  spe,
  reference,
  method = "wsum",
  assay = "counts",
  statistic = NULL,
  ...
)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{reference}{decoupleR reference}

\item{method}{calculation method to use}

\item{assay}{which assay to use}

\item{statistic}{select a sub results in case methods produce mutliple ones}

\item{...}{further arguments passed to the methods}
}
\description{
Compute Pathway activities with decoupleR
}
