% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{compute_forbidden_genes}
\alias{compute_forbidden_genes}
\title{Compute forbidden genes}
\usage{
compute_forbidden_genes(
  clean,
  suspect_gene_names = "",
  suspect_gene_patterns = "",
  seed = 123456
)
}
\arguments{
\item{clean}{anndata object}

\item{suspect_gene_names}{test}

\item{suspect_gene_patterns}{test}

\item{seed}{seed}
}
\description{
Compute forbidden genes
}
