% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{compute_metacells}
\alias{compute_metacells}
\title{Compute Metacells}
\usage{
compute_metacells(clean, forbidden_gene_names)
}
\arguments{
\item{clean}{anndata object}

\item{forbidden_gene_names}{list of genes names that are not used as metacell base}
}
\description{
Compute Metacells
}
