% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{compute_metacells}
\alias{compute_metacells}
\title{Compute Metacells}
\usage{
compute_metacells(
  clean,
  forbidden_gene_names,
  cell_type_col,
  target_size = 160000,
  abundance_score = 0.9,
  n_cores = NULL,
  seed = 12345
)
}
\arguments{
\item{clean}{anndata object}

\item{forbidden_gene_names}{list of genes names that are not used as metacell base}

\item{cell_type_col}{cell type column of cleaned anndata, used for reannotation}

\item{target_size}{target UMI count per metacells}

\item{abundance_score}{metacell celltype purity score}

\item{n_cores}{number of cores to use}

\item{seed}{seed}
}
\description{
Compute Metacells
}
