% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_conversion.R
\name{convert_human_to_mouse}
\alias{convert_human_to_mouse}
\title{map human to mouse genes}
\usage{
convert_human_to_mouse(humangenes)
}
\arguments{
\item{humangenes}{a vector containing human genes to be converted to mouse}
}
\description{
map human to mouse genes
}
