% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_conversion.R
\name{convert_mouse_to_human}
\alias{convert_mouse_to_human}
\title{map mouse to human genes}
\usage{
convert_mouse_to_human(mousegenes)
}
\arguments{
\item{mousegenes}{a vector containing mouse genes to be converted to human}
}
\description{
map mouse to human genes
}
