% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_expression.R
\name{corr_expr}
\alias{corr_expr}
\title{Heatmap correlation of cell types based on gene expression values}
\usage{
corr_expr(
  sig,
  log = FALSE,
  cor_method = c("pearson", "spearman"),
  matrix = FALSE
)
}
\arguments{
\item{sig}{signature matrix}

\item{log}{logarithmic transformation of signature matrix}

\item{cor_method}{correlation method "pearson" or "spearman"}

\item{matrix}{return correlation and correlation probability values as matrix}
}
\value{
correlation heatmap of cell types based on gene expression
}
\description{
Generates a heatmap of cell type correlations and p-values as asterisks based on signature matrix
}
