% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_algorithms.R
\name{deconvolute}
\alias{deconvolute}
\title{Deconvolution}
\usage{
deconvolute(
  spatial_object,
  single_cell_object,
  cell_type_col = "cell_ontology_class",
  method = NULL,
  ...
)
}
\arguments{
\item{spatial_object}{A SpatialExperiment}

\item{single_cell_object}{A SingleCellExperiment}

\item{cell_type_col}{Column name of the single_cell_object where the cell type can be found}

\item{method}{Deconvolution Method to use, see deconvolution_methods() for a full list of available methods}

\item{...}{Further parameters passed to the selected deconvolution method}
}
\value{
The deconvolution result as a table
}
\description{
Deconvolution
}
