% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_algorithms.R
\name{deconvolute}
\alias{deconvolute}
\title{Deconvolution with SpaceDeconv}
\usage{
deconvolute(
  spatial_obj,
  signature = NULL,
  single_cell_obj = NULL,
  cell_type_col = "cell_ontology_class",
  method = NULL,
  batch_id_col = NULL,
  assay_sc = "counts",
  assay_sp = "counts",
  return_object = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{spatial_obj}{A SpatialExperiment}

\item{signature}{Gene Expression Signature}

\item{single_cell_obj}{A SingleCellExperiment}

\item{cell_type_col}{Column name of the single_cell_obj where the cell type can be found}

\item{method}{Deconvolution Method to use, see deconvolution_methods() for a full list of available methods}

\item{batch_id_col}{column name of batch ids in single cell object}

\item{assay_sc}{which single cell assay to use for computation}

\item{assay_sp}{which spatial assay to use for computation}

\item{return_object}{Return an Object or result Table, TRUE = Object}

\item{verbose}{display more detailed information}

\item{...}{Further parameters passed to the selected deconvolution method}
}
\value{
The deconvolution result as a table
}
\description{
Perform cell type deconvolution with SpaceDeconv.
}
\examples{
# more examples can be found in the documentation website
data("single_cell_data_1")
data("spatial_data_1")

single_cell_data_1 <- SpaceDeconv::normalize(single_cell_data_1, method="cpm")

signature <- SpaceDeconv::build_model(
  single_cell_data_1,
  method="scdc",
  cell_type_column ="celltype_major",
  assay_sc="cpm",
  spatial_obj = spatial_data_1,
  assay_sp = "cpm",
  batch_id_col = "orig.ident"
)

deconvolution <- SpaceDeconv::deconvolute(
  spatial_obj = spatial_data_1,
  signature = signature,
  method = "scdc",
  single_cell_obj = single_cell_data_1,
  cell_type_col = "celltype_major",
  batch_id_col = "orig.ident"
)
}
