% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD.R
\name{deconvolute_card}
\alias{deconvolute_card}
\title{Deconvolute CARD}
\usage{
deconvolute_card(
  sce,
  spe,
  cell_type_col = "cell_ontology_class",
  assay_sc = "counts",
  assay_sp = "counts",
  batch_id_col = NULL
)
}
\arguments{
\item{sce}{Single Cell Experiment}

\item{spe}{Spatial Experiment}

\item{cell_type_col}{column of SCE containing cell type information}

\item{assay_sc}{assay of sce to use}

\item{assay_sp}{assay of spe to use}

\item{batch_id_col}{batch id column in spatialExperiment}
}
\description{
Deconvolute CARD
}
