% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD.R
\name{deconvolute_card}
\alias{deconvolute_card}
\title{Deconvolute CARD}
\usage{
deconvolute_card(
  single_cell_obj,
  spatial_obj,
  cell_type_col = "cell_ontology_class",
  assay_sc = "counts",
  assay_sp = "counts",
  batch_id_col = NULL,
  result_name = "card"
)
}
\arguments{
\item{single_cell_obj}{Single Cell Experiment}

\item{spatial_obj}{Spatial Experiment}

\item{cell_type_col}{column of SCE containing cell type information}

\item{assay_sc}{assay of single_cell_obj to use}

\item{assay_sp}{assay of spatial_obj to use}

\item{batch_id_col}{batch id column in spatialExperiment}

\item{result_name}{token to identify deconvolution results in object, default = "card"}
}
\description{
Deconvolute CARD
}
