% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOT.R
\name{deconvolute_dot}
\alias{deconvolute_dot}
\title{Deconvolute DOT}
\usage{
deconvolute_dot(
  single_cell_obj,
  spatial_obj,
  cell_type_col = "cell_ontology_class",
  assay_sc = "counts",
  assay_sp = "counts",
  result_name = "dot",
  ratios_weight = 0,
  max_spot_size = 20,
  ...
)
}
\arguments{
\item{single_cell_obj}{Single Cell Experiment}

\item{spatial_obj}{Spatial Experiment}

\item{cell_type_col}{column of SCE containing cell type information}

\item{assay_sc}{assay of single_cell_obj to use}

\item{assay_sp}{assay of spatial_obj to use}

\item{result_name}{token to identify deconvolution results in object, default = "dot"}

\item{ratios_weight}{This parameter determines how much the cell type ratios in the reference sc dataset are expected to match the ratios in the target spatial data. A higher value would match the cell type ratios more closely. By default this is 0 since we do not assume that the modalities match, but usually a value between 0.1 and 0.25 is reasonable depending on the quality of the sc data. If the two modalities are expected to match perfectly (e.g., come from the same tissue), a value closer to 1 would give the best results.}

\item{max_spot_size}{this parameter determines the maximum number of cells per spot and a higher value indicates lower resolution in the spatial data. The results should be robust for the default value but will give better estimates of the absolute abundances if the appropriate parameter is passed. For Visium, this is set to 20 (default). For ST it should be set to a higher value (e.g., 200).}

\item{...}{additional parameters passed to DOTr methods}
}
\description{
Deconvolute DOT
}
