% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOT.R
\name{deconvolute_dot}
\alias{deconvolute_dot}
\title{Deconvolute DOT}
\usage{
deconvolute_dot(
  single_cell_obj,
  spatial_obj,
  cell_type_col = "cell_ontology_class",
  assay_sc = "counts",
  assay_sp = "counts",
  result_name = "dot",
  ...
)
}
\arguments{
\item{single_cell_obj}{Single Cell Experiment}

\item{spatial_obj}{Spatial Experiment}

\item{cell_type_col}{column of SCE containing cell type information}

\item{assay_sc}{assay of single_cell_obj to use}

\item{assay_sp}{assay of spatial_obj to use}

\item{result_name}{token to identify deconvolution results in object, default = "dot"}

\item{...}{additional parameters passed to DOTr methods}
}
\description{
Deconvolute DOT
}
