% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunedeconv.R
\name{deconvolute_immunedeconv}
\alias{deconvolute_immunedeconv}
\title{Deconvolute Immundeconv}
\usage{
deconvolute_immunedeconv(spe, method = NULL, assay_sp = "counts", ...)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{method}{deconvolution algorithm}

\item{assay_sp}{assay of spatial object to use}

\item{...}{further parameters passed to the selected method}
}
\description{
Deconvolute Immundeconv
}
