% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunedeconv.R
\name{deconvolute_immunedeconv_mouse}
\alias{deconvolute_immunedeconv_mouse}
\title{Deconvolute Immunedeconv mouse}
\usage{
deconvolute_immunedeconv_mouse(
  spe,
  method = NULL,
  rmgenes = NULL,
  algorithm = NULL,
  assay_sp = "counts",
  ...
)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{method}{deconvolution algorithm}

\item{rmgenes}{genes to remove from the analysis}

\item{algorithm}{statistical algorithm for SeqImmuCC (ignored by all other methods)}

\item{assay_sp}{assay of spatial object to use}

\item{...}{additional parameters passed to function}
}
\description{
Deconvolute Immunedeconv mouse
}
