% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunedeconv.R
\name{deconvolute_immunedeconv_mouse}
\alias{deconvolute_immunedeconv_mouse}
\title{Deconvolute Immunedeconv mouse}
\usage{
deconvolute_immunedeconv_mouse(
  spatial_obj,
  method = NULL,
  rmgenes = NULL,
  algorithm = NULL,
  assay_sp = "counts",
  result_name = NULL,
  ...
)
}
\arguments{
\item{spatial_obj}{SpatialExperiment}

\item{method}{deconvolution algorithm}

\item{rmgenes}{genes to remove from the analysis}

\item{algorithm}{statistical algorithm for SeqImmuCC (ignored by all other methods)}

\item{assay_sp}{assay of spatial object to use}

\item{result_name}{token to identify deconvolution results in object, default = deconvolution method}

\item{...}{additional parameters passed to function}
}
\description{
Deconvolute Immunedeconv mouse
}
