% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute_omnideconv.R
\name{deconvolute_omnideconv}
\alias{deconvolute_omnideconv}
\title{Deconvolute Omnideconv}
\usage{
deconvolute_omnideconv(
  bulk_gene_expression,
  signature,
  method = NULL,
  single_cell_object = NULL,
  cell_type_annotations = NULL,
  batch_ids = NULL,
  cell_type_column_name = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{bulk_gene_expression}{A matrix or dataframe with the bulk data. Rows are genes, columns are spots.}

\item{signature}{A signature matrix}

\item{method}{omnideconv deconvolution method, see omnideconv::deconvolution_methods() for details}

\item{single_cell_object}{Needed for MOMF and Bisque, either a matrix with cell type annotations in the cell_type_annotations parameter or a singleCellExperiment / AnnData where the cell_type_column_name specifies the corresponding column where the cell type annotation is located in the object}

\item{cell_type_annotations}{Vector of cell type annotations, required if single_cell_object is a matrix}

\item{batch_ids}{A vector of the ids of samples or individuals}

\item{cell_type_column_name}{Required if single_cell_object is a SingleCellExperiment or Anndata. Specifing the column name where the cell type annotation can be found}

\item{verbose}{Whether to produce output on the console}

\item{...}{Additional parameters, passed to the selected method}
}
\description{
Deconvolute Omnideconv
}
