% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omnideconv.R
\name{deconvolute_omnideconv}
\alias{deconvolute_omnideconv}
\title{Deconvolute Omnideconv}
\usage{
deconvolute_omnideconv(
  spe,
  signature = NULL,
  method = NULL,
  single_cell_obj = NULL,
  cell_type_col = "cell_ontology_class",
  batch_id_col = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{signature}{A signature matrix}

\item{method}{omnideconv deconvolution method, see omnideconv::deconvolution_methods() for details}

\item{single_cell_obj}{Needed for MOMF and Bisque, either a matrix with cell type annotations in the cell_type_annotations parameter or a singleCellExperiment / AnnData where the cell_type_column_name specifies the corresponding column where the cell type annotation is located in the object}

\item{cell_type_col}{column name of single_cell_obj where cell type information is found}

\item{batch_id_col}{column name of single_cell_obj where batch ids can be found}

\item{verbose}{Whether to produce output on the console}

\item{...}{Additional parameters, passed to the selected method}
}
\description{
Deconvolute Omnideconv
}
