% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTD.R
\name{deconvolute_rctd}
\alias{deconvolute_rctd}
\title{RCTD Deconvolution}
\usage{
deconvolute_rctd(
  sc_counts,
  cell_types,
  n_umi_sc = NULL,
  spatial_counts,
  spatial_coords,
  n_umi_sp = NULL,
  n_cores = 20
)
}
\arguments{
\item{sc_counts}{(dgC) matrix of single cell counts}

\item{cell_types}{Named factor of cell types for the sc_counts matrix}

\item{n_umi_sc}{(optional) named list of umi counts for each cell}

\item{spatial_counts}{(dgC) matrix of spatial counts}

\item{spatial_coords}{data frame containing spatial coordinates}

\item{n_umi_sp}{(optional) named list of umi counts for each spot}

\item{n_cores}{Number of CPU cores to use for the calculation}
}
\description{
RCTD Deconvolution
}
