% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTD.R
\name{deconvolute_rctd}
\alias{deconvolute_rctd}
\title{RCTD Deconvolution}
\usage{
deconvolute_rctd(
  sce,
  cell_type_col,
  spe,
  n_umi_sc = NULL,
  n_umi_sp = NULL,
  n_cores = NULL
)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{cell_type_col}{Column containting cell type annotation}

\item{spe}{SpatialExperiment}

\item{n_umi_sc}{(optional) named list of umi counts for each cell}

\item{n_umi_sp}{(optional) named list of umi counts for each spot}

\item{n_cores}{Number of CPU cores to use for the calculation, NULL = use all cores}
}
\description{
RCTD Deconvolution
}
