% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTD.R
\name{deconvolute_rctd}
\alias{deconvolute_rctd}
\title{RCTD Deconvolution}
\usage{
deconvolute_rctd(
  single_cell_obj,
  cell_type_col,
  spatial_obj,
  assay_sc = "counts",
  assay_sp = "counts",
  n_umi_sc = NULL,
  n_umi_sp = NULL,
  n_cores = NULL
)
}
\arguments{
\item{single_cell_obj}{SingleCellExperiment}

\item{cell_type_col}{Column containting cell type annotation}

\item{spatial_obj}{SpatialExperiment}

\item{assay_sc}{single cell assay to use}

\item{assay_sp}{spatial assay to use}

\item{n_umi_sc}{(optional) named list of umi counts for each cell}

\item{n_umi_sp}{(optional) named list of umi counts for each spot}

\item{n_cores}{Number of CPU cores to use for the calculation, NULL = use all cores}
}
\description{
RCTD Deconvolution
}
