% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialDWLS.R
\name{deconvolute_spatial_dwls}
\alias{deconvolute_spatial_dwls}
\title{Deconvolute Spatial DWLS}
\usage{
deconvolute_spatial_dwls(spatial_obj, signature, assay_sp = "counts", ...)
}
\arguments{
\item{spatial_obj}{Spatial Experiment}

\item{signature}{Signature}

\item{assay_sp}{Assay of SpatialExperiment to use}

\item{...}{additional parameters}
}
\description{
Deconvolute Spatial DWLS
}
