% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPOTlight.R
\name{deconvolute_spotlight}
\alias{deconvolute_spotlight}
\title{Deconvolute SPOTlight}
\usage{
deconvolute_spotlight(spe, model = NULL)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{model}{SPOTlight Model}
}
\description{
Deconvolute SPOTlight
}
