% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_algorithms.R
\docType{data}
\name{deconvolution_methods}
\alias{deconvolution_methods}
\title{List of supported deconvolution methods}
\format{
An object of class \code{character} of length 32.
}
\usage{
deconvolution_methods
}
\description{
List of supported deconvolution methods
}
\details{
Spatial Algorithms \cr
\code{RCTD}, \code{SPOTlight}, \code{CARD}, \code{spatialDWLS}, \code{Cell2Location}, \code{DOT}

First-generation Methods (Immunedeconv) \cr
\code{MCPcounter}, \code{EPIC}, \code{quanTIseq}, \code{xCell}, \code{CIBERSORT}, \code{CIBERSORT (abs.)},
\code{TIMER}, \code{ConsensusTME}, \code{ABIS}, \code{ESTIMATE}

First-generation Mouse Methods (Immunedeconv) \cr
\code{mMCPcounter}, \code{seqImmuCC}, \code{DCQ}, \code{BASE}

This object is a named vector. The names correspond to the display name of the method,
the values to the internal name.
}
\keyword{datasets}
