% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_algorithms.R
\docType{data}
\name{deconvolution_methods}
\alias{deconvolution_methods}
\title{List of supported deconvolution methods}
\format{
An object of class \code{character} of length 32.
}
\usage{
deconvolution_methods
}
\description{
List of supported deconvolution methods
}
\keyword{datasets}
