% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{filter_sample_id}
\alias{filter_sample_id}
\title{Filter SPE to contain only one sample ID}
\usage{
filter_sample_id(spe, sample_id)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{sample_id}{sample_id}
}
\description{
Filter SPE to contain only one sample ID
}
