% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set.R
\name{gene_set_score}
\alias{gene_set_score}
\title{Calculate Gene Set Score}
\usage{
gene_set_score(spe, genes = NULL, assay = "cpm", name = "geneSet")
}
\arguments{
\item{spe}{SpatialExperiment object}

\item{genes}{vectors of genes to calculate gene set score}

\item{assay}{name of the assay to use, default is "cpm"}

\item{name}{name of the result column}
}
\value{
updated SpatialExperiment object including the gene set score
}
\description{
Calculate Gene Set Score
}
