% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{get_cluster_features}
\alias{get_cluster_features}
\title{Get cluster features}
\usage{
get_cluster_features(
  spe,
  clusterid = NULL,
  topn = 3,
  spmethod = NULL,
  zscore = TRUE
)
}
\arguments{
\item{spe}{spatialExperiment with cluster results}

\item{clusterid}{= name of the column with the clustering results}

\item{topn}{number of top features to be shown}

\item{spmethod}{spatial method used for the clustering, must be dorothea, collectri, progeny, expression or the name of the deconvolution method used}

\item{zscore}{= if the results should be z-score scaled or not}
}
\description{
Get cluster features
}
