% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR.R
\name{get_decoupleR_reference}
\alias{get_decoupleR_reference}
\title{Fetch Pathway or Transcription Factor Signatures from decoupleR}
\usage{
get_decoupleR_reference(
  method = "progeny",
  organism = "human",
  n_genes = 500,
  confidence = NULL,
  ...
)
}
\arguments{
\item{method}{The computational method to use for obtaining the reference: "progeny", "dorothea" or "collectri}

\item{organism}{The target organism for the reference signatures ("human", "mouse", etc.).}

\item{n_genes}{The number of top genes to return for Progeny pathways. Only applicable if \code{method} is "progeny".}

\item{confidence}{A vector of confidence levels (A, B, C, D, E) to filter the transcription factor targets from DoRothEA.
Only applicable if \code{method} is "dorothea".}

\item{...}{Additional parameters to fine-tune the method's execution.}
}
\value{
A data frame containing the requested reference
}
\description{
This function retrieves references for Progeny, DoRothEA or CollecTRI from the \code{decoupleR} package.
}
