% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPOTlight.R
\name{getMarkersSPOTlight}
\alias{getMarkersSPOTlight}
\title{Calculate Markers}
\usage{
getMarkersSPOTlight(single_cell_obj, cell_type_col = "cell_ontology_class")
}
\arguments{
\item{single_cell_obj}{SingleCellExperiment}

\item{cell_type_col}{Column containing the cell type

This Procedure reflects the suggestions of the SPOTlight authors, however,
they also state that there are other ways to calculate markers}
}
\description{
Calculate Markers
}
