% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{make_baseplot}
\alias{make_baseplot}
\title{Render spatial hex plot}
\usage{
make_baseplot(
  spe,
  df,
  to_plot,
  sample_id = "sample01",
  image_id = "lowres",
  show_image = TRUE,
  discrete = FALSE,
  offset_rotation = FALSE
)
}
\arguments{
\item{spe}{SpatialExperiment with deconvolution results}

\item{df}{containing the annotation to be plotted}

\item{to_plot}{column of df to plot}

\item{sample_id}{sample of the SpatialExperiment to be plotted}

\item{image_id}{image id for background image}

\item{show_image}{whether to show the spatial image}

\item{discrete}{should the color scale be discrete? Defaut = FALSE}

\item{offset_rotation}{correct hex orientation for rotated visium image}
}
\description{
Render spatial hex plot
}
