% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{make_baseplot}
\alias{make_baseplot}
\title{Render spatial hex plot}
\usage{
make_baseplot(
  spe,
  df,
  to_plot,
  palette = "Mako",
  transform_scale = NULL,
  sample_id = "sample01",
  reverse_palette = FALSE,
  image_id = "lowres",
  show_image = TRUE,
  palette_type = "sequential",
  offset_rotation = FALSE,
  spot_size = 1.17,
  limits = NULL,
  smooth = FALSE,
  smoothing_factor = 1.5,
  title_size = 30,
  font_size = 20,
  legend_size = 40,
  density = TRUE
)
}
\arguments{
\item{spe}{SpatialExperiment with deconvolution results}

\item{df}{containing the annotation to be plotted}

\item{to_plot}{column of df to plot}

\item{palette}{colorspace palette (sequential)}

\item{transform_scale}{data transform_scaleation to use, "log"}

\item{sample_id}{sample of the SpatialExperiment to be plotted}

\item{reverse_palette}{reverse color palette}

\item{image_id}{image id for background image}

\item{show_image}{whether to show the spatial image}

\item{palette_type}{"discrete", "sequential", "diverging"}

\item{offset_rotation}{correct hex orientation for rotated visium image}

\item{spot_size}{increase (>1) or decrease (<1) the hex size}

\item{limits}{vector of color scale limits}

\item{smooth}{whether to smooth the plot}

\item{smoothing_factor}{kernel size factor (multiples of spot distance)}

\item{title_size}{font size of title}

\item{font_size}{font size of legend}

\item{legend_size}{legend size in points}

\item{density}{whether to display a density distribution next to the spatial plot}
}
\description{
Render spatial hex plot
}
