% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize}
\alias{normalize}
\title{Normalize Expression Object}
\usage{
normalize(object, method = "cpm", assay = "counts")
}
\arguments{
\item{object}{SingleCellExperiment}

\item{method}{normalization method, ("cpm", "logcpm")}

<<<<<<< HEAD
\item{assay}{which assay to use}
=======
\item{assay}{assay to use}
>>>>>>> 39680fed84b9d6642f666ae9841aec75309bbe2c
}
\description{
Normalize Expression Object
}
