% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize}
\alias{normalize}
\title{Normalize Expression Object}
\usage{
normalize(object, method = "cpm")
}
\arguments{
\item{object}{SingleCellExperiment}

\item{method}{normalization method, ("cpm")}
}
\description{
Normalize Expression Object
}
