% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_cells_per_spot}
\alias{plot_cells_per_spot}
\title{Plot Cells per Spot}
\usage{
plot_cells_per_spot(
  spatial_obj,
  plot_type = "spatial",
  threshold = 0.01,
  palette = "Rocket",
  transform_scale = NULL,
  reverse_palette = FALSE,
  sample_id = "sample01",
  image_id = "lowres",
  show_image = TRUE,
  offset_rotation = FALSE,
  spot_size = 1,
  limits = NULL
)
}
\arguments{
\item{spatial_obj}{SpatialExperiment containing deconvolution results}

\item{plot_type}{bar chart or spatial (bar, spatial)}

\item{threshold}{threshold for presence/absence, single value or vector of length nrow(spatial_obj)}

\item{palette}{colorspace palette (sequential)}

\item{transform_scale}{data transform_scaleation to use, "log"}

\item{reverse_palette}{reverse color palette}

\item{sample_id}{sample of SpatialExperiment to be plotted}

\item{image_id}{image of SpatialExperiment for the background annotation}

\item{show_image}{whether to show the histology image in the background}

\item{offset_rotation}{correct hex orientation for rotated visium image}

\item{spot_size}{increase (>1) or decrease (<1) the hex size}

\item{limits}{vector of color scale limit}
}
\value{
A hex plot containing unique cell counts per spot
}
\description{
Plot Cells per Spot
}
\examples{
data("spatial_data_1")
deconv <- SpaceDeconv::deconvolute(spatial_data_1, method = "estimate")
SpaceDeconv::plot_cells_per_spot(deconv)
}
