% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_cells_per_spot}
\alias{plot_cells_per_spot}
\title{Plot Cells per Spot}
\usage{
plot_cells_per_spot(
  spatial_obj,
  plot_type = "spatial",
  threshold = 0,
  spot_size = 1.5,
  show_image = TRUE
)
}
\arguments{
\item{spatial_obj}{SpatialExperiment containing deconvolution results}

\item{plot_type}{bar chart or spatial (bar, spatial)}

\item{threshold}{threshold for presence/absence, single value or vector of length nrow(spatial_obj)}

\item{spot_size}{size of the dots}

\item{show_image}{whether to show the histology image in the background}
}
\description{
Plot Cells per Spot
}
