% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_celltype}
\alias{plot_celltype}
\title{Plot Spatial Object}
\usage{
plot_celltype(
  spatial_obj,
  sample = "sample01",
  cell_type = NULL,
  plot_density = TRUE,
  spot_size = 1,
  show_image = TRUE
)
}
\arguments{
\item{spatial_obj}{SpatialExperiment containing deconvolution results}

\item{sample}{sample to plot}

\item{cell_type}{Cell Type to plot}

\item{plot_density}{(default = FALSE) wheter to plot the density}

\item{spot_size}{size of the spots in the plot}

\item{show_image}{whether to show the histology image in the background}
}
\description{
Plot Spatial Object
}
