% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_celltype}
\alias{plot_celltype}
\title{Function to plot deconvolution results}
\usage{
plot_celltype(
  spe,
  cell_type = NULL,
  palette = "Rocket",
  transform_scale = NULL,
  sample_id = "sample01",
  image_id = "lowres",
  reverse_palette = FALSE,
  show_image = TRUE,
  discrete = FALSE,
  offset_rotation = FALSE,
  spot_size = 1,
  limits = NULL,
  smooth = FALSE,
  smoothing_factor = 1.5,
  font_size = 15
)
}
\arguments{
\item{spe}{deconvolution result in Form of a SpatialExperiment}

\item{cell_type}{one or more celltype to plot}

\item{palette}{colorspace palette (sequential)}

\item{transform_scale}{data transform_scaleation to use, "log"}

\item{sample_id}{sample id to plot, default: "
sample01"}

\item{image_id}{which image to plot, default: "lowres"}

\item{reverse_palette}{reverse color palette}

\item{show_image}{logical, whether to display the image, default = TRUE}

\item{discrete}{logical, whether to scale the color discrete, default = FALSE}

\item{offset_rotation}{correct hex orientation for rotated visium image}

\item{spot_size}{increase (>1) or decrease (<1) the hex size}

\item{limits}{vector of color scale limits}

\item{smooth}{whether to smooth the plot}

\item{smoothing_factor}{kernel size factor (multiples of spot distance)}

\item{font_size}{font size}
}
\value{
plot of cell type fractions
}
\description{
Generate Hex Plot of a SpatialExperiment containing deconvolution results
}
\examples{
data("spatial_data_2")
deconv <- SpaceDeconv::deconvolute(spatial_data_2, method = "estimate")
SpaceDeconv::plot_celltype(deconv, cell_type = "estimate_immune.score")
}
