% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_celltype_presence}
\alias{plot_celltype_presence}
\title{Plot celltype presence absence}
\usage{
plot_celltype_presence(
  spe,
  cell_type = NULL,
  threshold = 0.01,
  palette = "Mako",
  transform_scale = NULL,
  sample_id = "sample01",
  image_id = "lowres",
  reverse_palette = FALSE,
  show_image = FALSE,
  offset_rotation = FALSE,
  spot_size = 1.17,
  limits = NULL,
  smooth = FALSE,
  smoothing_factor = 1.5,
  title_size = 30,
  font_size = 20,
  legend_size = 40
)
}
\arguments{
\item{spe}{deconvolution result in Form of a SpatialExperiment}

\item{cell_type}{celltype to plot}

\item{threshold}{fraction threshold, everything above is counted as "detected"}

\item{palette}{colorspace palette (sequential)}

\item{transform_scale}{data transform_scaleation to use, "log"}

\item{sample_id}{sample id to plot, default: "sample01"}

\item{image_id}{which image to plot, default: "lowres"}

\item{reverse_palette}{reverse color palette}

\item{show_image}{logical, wether to display the image, default = TRUE}

\item{offset_rotation}{correct hex orientation for rotated visium image}

\item{spot_size}{increase (>1) or decrease (<1) the hex size}

\item{limits}{vector of color scale limits}

\item{smooth}{whether to smooth the plot}

\item{smoothing_factor}{kernel size factor (multiples of spot distance)}

\item{title_size}{font size of title}

\item{font_size}{font size of legend}

\item{legend_size}{legend size in points}
}
\value{
plot of a celltypes presence/absence using a threshold
}
\description{
Plot celltype presence absence
}
