% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripleys.R
\name{plot_ripleys_k}
\alias{plot_ripleys_k}
\title{Combined Ripley´s K statistics in one plot}
\usage{
plot_ripleys_k(k_functions)
}
\arguments{
\item{k_functions}{named list of k statistics from different SpatialExperiments}
}
\value{
combined Ripley´s K results in one plot
}
\description{
Combined Ripley´s K statistics in one plot
}
