% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Benchmarking scatterplot to compare two spatial objects}
\usage{
plot_scatter(spe1, value1, spe2, value2, log_scale = FALSE)
}
\arguments{
\item{spe1}{SpatialExperiment}

\item{value1}{deconvolution result to plot}

\item{spe2}{SpatialExperiment}

\item{value2}{deconvolution result to plot}

\item{log_scale}{whether to log scale the axes}
}
\description{
Benchmarking scatterplot to compare two spatial objects
}
