% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_umi_count}
\alias{plot_umi_count}
\title{Function to plot deconvolution results}
\usage{
plot_umi_count(
  spe,
  sample_id = "sample01",
  image_id = "lowres",
  show_image = TRUE
)
}
\arguments{
\item{spe}{deconvolution result in Form of a SpatialExperiment}

\item{sample_id}{sample id to plot, default: "sample01"}

\item{image_id}{which image to plot, default: "lowres"}

\item{show_image}{logical, wether to display the image, default = TRUE}
}
\value{
plot of cell type fractions
}
\description{
Generate Hex Plot of a SpatialExperiment containing UMI counts
}
\examples{
data("spatial_data_3")
deconv <- SpaceDeconv::deconvolute(spatial_data_3, method="estimate")
plot_umi_count(deconv)
}
