% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_umi_count}
\alias{plot_umi_count}
\title{Function to plot deconvolution results}
\usage{
plot_umi_count(
  spe,
  palette = "Rocket",
  transform_scale = NULL,
  sample_id = "sample01",
  image_id = "lowres",
  reverse_palette = FALSE,
  show_image = TRUE,
  offset_rotation = FALSE,
  spot_size = 1,
  limits = NULL,
  smooth = FALSE,
  smoothing_factor = 1.5
)
}
\arguments{
\item{spe}{deconvolution result in Form of a SpatialExperiment}

\item{palette}{colorspace palette (sequential)}

\item{transform_scale}{data transform_scaleation to use, "log"}

\item{sample_id}{sample id to plot, default: "sample01"}

\item{image_id}{which image to plot, default: "lowres"}

\item{reverse_palette}{reverse color palette}

\item{show_image}{logical, wether to display the image, default = TRUE}

\item{offset_rotation}{correct hex orientation for rotated visium image}

\item{spot_size}{increase (>1) or decrease (<1) the hex size}

\item{limits}{vector of color scale limits}

\item{smooth}{whether to smooth the plot}

\item{smoothing_factor}{kernel size factor (multiples of spot distance)}
}
\value{
plot of cell type fractions
}
\description{
Generate Hex Plot of a SpatialExperiment containing UMI counts
}
\examples{
data("spatial_data_3")
deconv <- SpaceDeconv::deconvolute(spatial_data_3, method = "estimate")
plot_umi_count(deconv)
}
