% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_info.R
\name{print_info}
\alias{print_info}
\title{Print info about dataset}
\usage{
print_info(sce = NULL, spe = NULL, signature = NULL)
}
\arguments{
\item{sce}{singleCellExperiment}

\item{spe}{SpatialExperiment}

\item{signature}{signature matrix}
}
\description{
Print info about dataset
}
