% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualitycontrol.R
\name{qualitycontrol}
\alias{qualitycontrol}
\title{Quality control function}
\usage{
qualitycontrol(spe, assay = "counts")
}
\arguments{
\item{spe}{SpatialExperiment object}

\item{assay}{assay}
}
\value{
html file with quality control metrics
}
\description{
Generates a html report containing quality metrics of the SpatialExperiment
}
