% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripleys.R
\name{ripleys_k}
\alias{ripleys_k}
\title{Ripley´s K celltype distribution}
\usage{
ripleys_k(spe, cell_type, method, threshold = NULL)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{cell_type}{celltype of interest}

\item{method}{deconvolution method}

\item{threshold}{cutoff for cell type presence}
}
\value{
ripley´s k statistics
}
\description{
Ripley´s K celltype distribution
}
