% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_cell_counts.R
\name{scale_cell_counts}
\alias{scale_cell_counts}
\title{Scale relative cell fractions to absolute by providing cell densities}
\usage{
scale_cell_counts(spe, value, cell_counts, resName = NULL)
}
\arguments{
\item{spe}{An SpatialExperiment object containing spatial transcriptomics data.}

\item{value}{A string specifying the name of the column in \code{colData(spe)} that contains
relative cell fractions/deconvolution results to be scaled.}

\item{resName}{Optional; a string specifying the name of the new column to be added
to \code{colData(spe)} that will contain the absolute cell densities.
If NULL, defaults to appending "_absolute" to the \code{value} parameter.}

\item{cell_densities}{A named vector of absolute cell densities, where names correspond
to the rownames in \code{colData(spe)}.}
}
\value{
SPE object with added absolute celltype counts for each spot
}
\description{
This function calculates absolute cell densities for a SpatialExperiment object
by scaling a specified column of relative cell fractions with provided absolute cell densities.
The result is added as a new column in the \code{colData} of the SPE object.
}
