% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_correlation.R
\name{spatialcorr}
\alias{spatialcorr}
\title{create a correlation plot}
\usage{
spatialcorr(spe, method, adjust = "fdr", variables = NULL, sig.level = 0.05)
}
\arguments{
\item{spe}{#the spe object}

\item{method}{#method to used for correlation analysis - deconvolution method or decoupleR e.g. "cell2location"}

\item{adjust}{#method used to adjust p-values for multiple testing}

\item{variables}{#if not provided, the function selects variables based on the specified method, it can also be a vector}

\item{sig.level}{#it can also be a vector}
}
\description{
create a correlation plot
}
