% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_correlation.R
\name{spatialcorr}
\alias{spatialcorr}
\title{create a correlation plot}
\usage{
spatialcorr(
  spe,
  method,
  adjust = "fdr",
  variables = NULL,
  sig.level = 0.05,
  type = "lower",
  diag = FALSE,
  order = "original",
  plot_layout = "color",
  insig = "label_sig",
  addCoef.col = NULL,
  ...
)
}
\arguments{
\item{spe}{the spe object}

\item{method}{method to used for correlation analysis - deconvolution method or decoupleR e.g. "cell2location"}

\item{adjust}{method used to adjust p-values for multiple testing, see stats::p.asjust.methods for details}

\item{variables}{if not provided, the function selects variables based on the specified method, it can also be a vector}

\item{sig.level}{it can also be a vector}

\item{type}{Character, 'full' 'upper' or 'lower' (default), display full matrix, lower triangular or upper triangular matrix.}

\item{diag}{Logical, whether display the correlation coefficients on the principal diagonal. default is FALSE}

\item{order}{Character, the ordering method of the correlation matrix.'original' for original order (default),'AOE' for the angular order of the eigenvectors, 'FPC' for the first principal component order, 'hclust' for the hierarchical clustering order,'alphabet' for alphabetical order.}

\item{plot_layout}{represents the method parameter in the original corrplot function.The layout of the correlation plot. color is default. Choose between "circle", "square", "ellipse", "number", "shade", "color", "pie"}

\item{insig}{If 'blank', wipe away the corresponding glyphs; if 'p-value', add p-values the corresponding glyphs; if 'pch', add characters (see pch for details) on corresponding glyphs; if 'n', don't take any measures; if 'label_sig' (default), mark significant correlations with pch (see sig.level).}

\item{addCoef.col}{to add labels to the plot showing the correlation values, default is NULL, the user can choose any other color to add and color the values.}

\item{...}{additional parameters passed to corrplot function}
}
\description{
create a correlation plot
}
