% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subsetSCE}
\alias{subsetSCE}
\title{Subset Single Cell Experiment}
\usage{
subsetSCE(
  sce,
  cell_type_col = "celltype_major",
  scenario = "even",
  ncells = 1000,
  notEnough = "asis",
  seed = 12345
)
}
\arguments{
\item{sce}{SingleCellExperiment with cell type annotation}

\item{cell_type_col}{name of colData column containing cell type information}

\item{scenario}{which fraction of each celltype should be present in the subsampled dataset?}

\item{ncells}{number of cells in total, defaults to 1000}

\item{notEnough}{what to do if not enough cells are available c("remove", "asis")}

\item{seed}{set a seed, enables reproducibility of subsetting step}
}
\value{
a singleCellExperiment with
}
\description{
Subset a SingleCellExperiment to reduce the number of cells
}
