% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subsetSCE}
\alias{subsetSCE}
\title{Subset a SingleCellExperiment by Cell Type}
\usage{
subsetSCE(
  sce,
  cell_type_col = "celltype_major",
  scenario = "even",
  ncells = 1000,
  notEnough = "asis",
  seed = 12345
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object containing cell type annotations. The function applies subsetting
operations on this object to achieve a targeted composition of cell types.}

\item{cell_type_col}{A character string indicating the column name within \code{colData} of the \code{SingleCellExperiment}
object that contains cell type information. This column is used to identify and categorize cells by type for subsetting.}

\item{scenario}{Specifies the strategy for cell selection across types: \code{"even"} for equal numbers from each cell
type, or \code{"mirror"} to reflect the proportional representation of cell types in the subset as in the original dataset.}

\item{ncells}{The target total number of cells for the subsetted object. Defaults to 1000. This parameter dictates
the overall size of the subsetted \code{SingleCellExperiment}.}

\item{notEnough}{Defines the behavior when there are not enough cells of a certain type to meet the subsetting criteria:
\code{"remove"} to exclude that cell type entirely, or \code{"asis"} to include all available cells of that type despite falling short
of the target number.}

\item{seed}{An optional numeric value to set the random seed for reproducibility of the subsetting process. This ensures
that the subset selection can be replicated in future analyses.}
}
\value{
Returns a subsetted \code{SingleCellExperiment} object. The resulting object reflects the targeted distribution and
number of cells as specified by the input parameters. It enables focused analysis on a more manageable dataset while
preserving the desired representation of cell types.
}
\description{
This function facilitates the subsetting of a \code{SingleCellExperiment} object to a specified number of cells,
with an option to maintain an even or proportional distribution across different cell types. It offers flexibility
in handling scenarios where insufficient cells of a specific type are available and ensures reproducibility through
the option to set a seed for random sampling.
}
