% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subsetSPE}
\alias{subsetSPE}
\title{Subset a SpatialExperiment Object}
\usage{
subsetSPE(spe, colRange = NULL, rowRange = NULL)
}
\arguments{
\item{spe}{A \code{SpatialExperiment} object containing spatial coordinates and associated data.}

\item{colRange}{A numeric vector of length 2, specifying the minimum and maximum pxl_col_in_fullres.}

\item{rowRange}{A numeric vector of length 2, specifying the minimum and maximum pxl_row_in_fullres.}
}
\value{
A subset of the \code{SpatialExperiment} object, including only the data within the specified coordinate ranges.
}
\description{
This function subsets a \code{SpatialExperiment} object based on specified x and y coordinate ranges.
It is useful in spatial transcriptomics for isolating specific regions of interest.
}
