% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subsetSPE}
\alias{subsetSPE}
\title{Isolate Regions within a SpatialExperiment Object}
\usage{
subsetSPE(spe, colRange = NULL, rowRange = NULL)
}
\arguments{
\item{spe}{The \code{SpatialExperiment} object representing spatially resolved transcriptomics data.}

\item{colRange}{a numeric vector of length two specifying the minimum and maximum x-coordinates
(pxl_col_in_fullres) to include in the subset. If \code{NULL}, the full range of x-coordinates is used.}

\item{rowRange}{a numeric vector of length two specifying the minimum and maximum y-coordinates
(pxl_row_in_fullres) to include in the subset. If \code{NULL}, the full range of y-coordinates is used.}
}
\value{
Returns a \code{SpatialExperiment} object subsetted to include only spots within the defined x and y coordinate ranges.
}
\description{
Tailor a \code{SpatialExperiment} object to focus on specific areas by defining ranges in the spatial coordinates.
This function narrows down the dataset to include only those spots that fall within user-defined x and y coordinate
boundaries.
}
